redirects

+++ secudevcase2/deletepost.php
@@ -2,6 +2,12 @@
 include 'session.php';
 include 'connect.php';
 
+if(!isset($_SESSION['signed_in']) || $_SESSION['signed_in'] == false)
+{
+	header("Location: /main.php"); /* Redirect browser */
+	exit();
+}
+
 $query = "DELETE FROM posts WHERE id = :id";
 
 try{
